Remodel waking
Instead of somehow tying the wakeup mechanism to the clock, we now only
support boolean-flag-style waking. In many cases, this isn't even going
to be used effectively, but instead we'll be polling on something.

In fact, it'll always be some kind of polling, because the executor loop
also just checks whether a future is ready to run.